introduce file info fragment, separate image detail logic
