+++ app/src/main/res/drawable/ic_dashboard_outlined.xml
@@ -1,0 +1,24 @@
+<!--
+  ~ Copyright (C) 2026 The Android Open Source Project
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M520,360v-240h320v240L520,360ZM120,520v-400h320v400L120,520ZM520,840v-400h320v400L520,840ZM120,840v-240h320v240L120,840ZM200,440h160v-240L200,200v240ZM600,760h160v-240L600,520v240ZM600,280h160v-80L600,200v80ZM200,760h160v-80L200,680v80ZM360,440ZM600,280ZM600,520ZM360,680Z"
+      android:fillColor="#000000"/>
+</vector>
