+++ app/src/main/res/layout/file_info_fragment.xml
